fix: create bucket_enabled flag required if bucket_id provided
Motivation:

Without this flag the local.s3_bucket_enabled flag is calculated based
on:
s3_bucket_enabled = var.bucket_id == ""

The problem with above is that when using this module, the terraform
(12.10) gives following error:

```
Error: Invalid count argument

  on .terraform/modules/s3-image-resizer.s3-bucket-api-images.s3_user.s3_user/main.tf line 14, in resource "aws_iam_user" "default":
  14:   count         = var.enabled ? 1 : 0

The "count" value depends on resource attributes that cannot be determined
until apply, so Terraform cannot predict how many instances will be created.
To work around this, use the -target argument to first apply only the
resources that the count depends on.
```

In order to help terraform with determination if resources should be
created or not we have to create this variable explicity.

Approved-by: Konrad Obal